Rewrite operator table.
The OpTable is now a sorted slice. The sort order was chosen carefully to
simultaneously meet the various order requirements of the lexer and parser,
removing the need to do sorting during parsing.

All of the operator stuff is now part of the parse package. This makes more
sense because operators are inherently coupled with the parser.

The new OpTable is not synchronized. This may or may not need to be done in the
future.

This reverts commit 648c1424738d658dbcfc15277d14667aef253ee1 (syncronization).
This reverts commit e1c5309a42be1ed72723121e054eb78dcf7bea3e (organization).