Rewrite the parser.
The parser now only buffers a single lexeme and no longer backtracks. Yay!

The lexer is rewritten, but works mostly the same. No kind of operator
identifying is present in the new lexer, which was a bad idea in the old one.

All of the types in lang/term are removed. Instead, the parser provides a single
Term type. By reducing all terms to a single type, the tree can be allocated
contiguously. Plus, the heap is built in bottom-up level-order, which is a
required traversal that was previously expensive.

The operator interface is basically the same but made more opaque. I'm not
settled on the current data structure, so this gives me room to work.