Introduce Values, Names, and Namespaces.
Value is the common interface of all atomic Prolog data types. To perform
unification on Terms, we must compare the Values of the terms. However, Values
are polymorphic (numbers, functors, variables, etc.) and cannot be stored within
the Term struct. Instead, we store Values in a Namespace. The Namespace provides
a Name for each of its Values. Names are embedded in Terms in place of Values,
and can be used to retrieve Values from the Namespace. Names carry enough
information to perform unification without querying the Namespace.