Refactor and fix bugs
I've been exploring the API design space, and I'm reasonably happy with
where we stand right now with the lang package. There is still some
work to be done with regards to synchronizing the lexer and parser.
Currently the lexer halts after scanning one clause, and the parser
only processes one clause at a time. In the future, I would like the
lexer to scan the entire input and the parser to produce a stream of
clauses. The parser will need to pause after directives to allow the
caller to update the operator table if necessary.