Rewrite lexer
I switched from Pike's technique used in text/template to using a
classic scanner. This more than doubled the performance of the lexer!

The symbol types are no longer coupled to the scanner. They will be
moved to the scope package, and the package will be renamed to
something more appropriate.