Introduce new addressing technique
Previously, a Namespace would generate an address for functors between
zero and one by computing the average of the available space. Because
of the way IEEE 754 floats work, this left very little entropy in the
name.

The new technique gives up on an upper bound and generates addresses
from a constant "base". Addresses generated past the existing
upper/lower edge of the tree are set to the address of the existing
edge multiplied/divided by the base. Because of the way float64s are
represented, a base of 2 results in an address space with very little
entropy. A base of 31 appears to have better entropy to support hashing
in the database.